Splits core.Color from api.Color for parser access to color literals
